tests/test_asm.sh added to run unit tests for compiler other minor fixes

+++ asm.sh
@@ -1,31 +1,32 @@
 #!/usr/bin/env bash
-# this script allows to convert .kgasm files to the machine code
+# this script allows to convert .kga files to the machine code
 
 LABELS=()
 LABELS_ADDRESSES=()
-LABELS_COUNT=0
 
 CONSTANTS=()
 CONSTANTS_EVALUATED=()
 CONSTANTS_ADDRESSES=()
-CONSTANTS_COUNT=0
 
 VARIABLES=()
 VARIABLES_ADDRESSES=()
 VARIABLES_DECL_ADDRESSES=()
-VARIABLES_COUNT=0
 
 COMPILATION_ERROR_COUNT=0
 
 CUR_FILE=
 CUR_LINE_NO=
-LINE=
+CUR_LINE=
+
+if [ -z "$DEBUG_INFO" ]; then
+    DEBUG_INFO=1
+fi
 
 function compilation_error() {
     local EXPECTED_SYNTAX="$1"
     local ERROR_INFO="$2"
     echo -e "\033[93mCompilation error\033[0m at ${CUR_FILE}:${CUR_LINE_NO}" 1>&2
-    echo -e "\033[91m$LINE\033[0m" 1>&2
+    echo -e "\033[91m$CUR_LINE\033[0m" 1>&2
     if [ -n "$ERROR_INFO" ]; then
         echo -e "\033[91m$ERROR_INFO\033[0m" 1>&2
     fi
@@ -33,25 +34,27 @@ function compilation_error() {
     echo 1>&2
 
     COMPILATION_ERROR_COUNT=$((COMPILATION_ERROR_COUNT + 1))
+    if [ $COMPILATION_ERROR_COUNT -gt 20 ]; then
+        echo "Too many compilation errors, aborting" 1>&2
+        exit 1
+    fi
 }
 
 function contains_element() {
-    local value="$1"  # Value to search for
-    shift             # Shift to access array arguments
-    local array=("$@") # Remaining arguments are the array
+    local value="$1"; shift
+    local array=("$@")
 
     for element in "${array[@]}"; do
         if [[ "$element" == "$value" ]]; then
-            return 0 # Found the value
+            return 0
         fi
     done
-    return 1 # Value not found
+    return 1
 }
 
 function find_index() {
-    local element="$1"  # The element to search for
-    shift               # Remove the first argument
-    local array=("$@")  # The rest are the array elements
+    local element="$1"; shift
+    local array=("$@")
 
     for i in "${!array[@]}"; do
         if [[ "${array[i]}" == "$element" ]]; then
@@ -75,12 +78,15 @@ function eval_constant_lexeme() {
     esac
 }
 
-source include/operations.sh
-source include/other.sh
-source include/registers.sh
-source include/system.sh
+if [ -z "$INCLUDE_DIR" ]; then
+    INCLUDE_DIR="$(dirname "$0")"/include
+fi
 
-# TODO More careful checks of errors - no files, permissions and so on
+source "$INCLUDE_DIR"/operations.sh
+source "$INCLUDE_DIR"/other.sh
+source "$INCLUDE_DIR"/registers.sh
+source "$INCLUDE_DIR"/system.sh
+
 KERNEL_FILE="$GLOBAL_KERNEL_DISK"
 SRC_FILES=""
 for ARG in "$@"; do
@@ -112,10 +118,8 @@ function is_constant_lexeme() {
         return 1;
     fi
     case $2 in
-    num|str|opr|clr)
-        return 0;;
-    *)
-        return 1;;
+    num|str|opr|clr)    return 0;;
+    *)                  return 1;;
     esac
 }
 
@@ -136,7 +140,7 @@ function parse_lexeme() {
     if [ "${CUR_LEXEME:0:1}" = '"' ] && [ "${CUR_LEXEME: -1}" = '"' ]; then
         echo "$PREFIX str ${CUR_LEXEME:1:-1}"
     elif [[ "$CUR_LEXEME" =~ ^[0-9]+$ ]]; then
-        echo "$PREFIX num"
+        echo "$PREFIX num $CUR_LEXEME"
     elif [ "$CUR_LEXEME" = "to" ]; then
         echo "$PREFIX key to"
     elif [ "${CUR_LEXEME:0:4}" = "var:" ]; then
@@ -173,16 +177,17 @@ PARSED_LEXEMES=()
 for CUR_FILE in $SRC_FILES; do
     PARSED_LEXEMES+=( "$(echo file ${CUR_FILE})" )
     CUR_LINE_NO=0
-    while read -r LINE; do
+
+    while read -r CUR_LINE || [ -n "$CUR_LINE" ]; do
         CUR_LINE_NO=$((CUR_LINE_NO + 1))
         # remove leading and trailing spaces
-        LINE=$(echo "${LINE}" | sed -e 's/^[[:space:]]*//' -e 's/[[:space:]]*$//')
-        if [ -z "$LINE" ] || [ "${LINE:0:2}" = "//" ]; then
+        CUR_LINE=$(echo "${CUR_LINE}" | sed -e 's/^[[:space:]]*//' -e 's/[[:space:]]*$//')
+        if [ -z "$CUR_LINE" ] || [ "${CUR_LINE:0:2}" = "//" ]; then
             continue
         fi
 
         # Parse line to components:
-        eval set -- "$LINE"
+        eval set -- "$CUR_LINE"
 
         # Let's set expected lexemes count depending on the command
         CUR_CMD="$1"
@@ -200,7 +205,7 @@ for CUR_FILE in $SRC_FILES; do
             ARG_IND=$((i+1))
             CUR_LEXEME="${!ARG_IND}"
             # eval command eliminates double quotes from string literals. Let's move them back for the valid case:
-            if [ "$i" -eq 1 ] && [[ $(echo "$LINE" | awk '{print $1" "$2}') == 'write "'* ]]; then
+            if [ "$i" -eq 1 ] && [[ $(echo "$CUR_LINE" | awk '{print $1" "$2}') == 'write "'* ]]; then
                 CUR_LEXEME="\"$CUR_LEXEME\""
             fi
 
@@ -293,23 +298,40 @@ done
 
 function get_address_for_lexeme() {
     local LEX="$1"
+    local PREFIX="${LEX:0:1}"
+    if [ "$PREFIX" = "_" ]; then
+        PREFIX=""
+    fi
     local TYPE="${LEX:2:3}"
     local VALUE="${LEX:6}"
+    FUNC_RESULT=""
     case $TYPE in
     reg)
-        eval echo "\$$VALUE";;
+        FUNC_RESULT=$(eval echo "${PREFIX}\$$VALUE");;
     num)
-        echo "$VALUE";;
+        FUNC_RESULT="${PREFIX}$VALUE";;
     lbl)
-        CUR_INDEX=$(find_index "$VALUE" "${LABELS[@]}")
-        echo "${LABELS_ADDRESSES[$CUR_INDEX]}";;
+        CUR_INDEX=$(find_index "${VALUE}" "${LABELS[@]}")
+        if [ -n "$PREFIX" ]; then
+            compilation_error "" "label:$VALUE should not have prefix $PREFIX"
+        elif [ "$CUR_INDEX" -eq -1 ]; then
+            compilation_error "" "Label $VALUE is not defined"
+        else
+            FUNC_RESULT="${LABELS_ADDRESSES[$CUR_INDEX]}"
+        fi
+        ;;
     var)
         CUR_INDEX=$(find_index "$VALUE" "${VARIABLES[@]}")
-        DECL_ADDRESS=${VARIABLES_DECL_ADDRESSES[$CUR_INDEX]}
-        if [ "$CUR_INSTRUCTION_NO" -lt "$DECL_ADDRESS" ]; then
-            compilation_error "Variable should be used after declaration" "Variable $VALUE is used before declaration"
+        if [ "$CUR_INDEX" -eq -1 ]; then
+            compilation_error "Variable should be declared before usage" "Variable $VALUE is not defined"
+        else
+            DECL_ADDRESS=${VARIABLES_DECL_ADDRESSES[$CUR_INDEX]}
+            if [ "$CUR_INSTRUCTION_NO" -lt "$DECL_ADDRESS" ]; then
+                compilation_error "Variable should be defined before the first use" "Variable ${VALUE} is used before declaration"
+            fi
+            FUNC_RESULT="${PREFIX}${VARIABLES_ADDRESSES[$CUR_INDEX]}"
         fi
-        echo "${VARIABLES_ADDRESSES[$CUR_INDEX]}";;
+        ;;
     *)  
         ;;
     esac
@@ -322,7 +344,7 @@ function get_debug_address_for_lexeme() {
     if [ "$PREFIX" = "_" ]; then
         PREFIX=""
     fi
-    echo "$PREFIX$VALUE"
+    FUNC_RESULT="$PREFIX$VALUE"
 }
 
 rm -rf "${KERNEL_FILE}"
@@ -357,20 +379,25 @@ while true; do
     CUR_ERROR=""
     CUR_CMD="${LEXEME:6}"
     RES_STR=""
+    DEBUG_STR=""
     case "$CUR_CMD" in
     cpu_exec)
-        RES_STR="$INSTR_CPU_EXEC # cpu_exec" ;;
+        RES_STR="$INSTR_CPU_EXEC"
+        DEBUG_STR="cpu_exec" ;;
     DEBUG_ON|DEBUG_OFF)
         RES_STR="$CUR_CMD" ;;
     jump|jump_if)
         get_next_lexeme
-        CUR_ADDR_TO=$(get_address_for_lexeme "${LEXEME}")
-        CUR_DEBUG_TO=$(get_debug_address_for_lexeme "${LEXEME}")
+        get_address_for_lexeme "${LEXEME}"
+        CUR_ADDR_TO="$FUNC_RESULT"
+        get_debug_address_for_lexeme "${LEXEME}"
+        CUR_DEBUG_TO="$FUNC_RESULT"
         if [ "$CUR_CMD" = jump ]; then
-            RES_STR="$INSTR_JUMP $CUR_ADDR_TO # jump $CUR_DEBUG_TO"
+            RES_STR="$INSTR_JUMP ${CUR_ADDR_TO}"
         else
-            RES_STR="$INSTR_JUMP_IF $CUR_ADDR_TO # jump_if $CUR_DEBUG_TO"
+            RES_STR="$INSTR_JUMP_IF ${CUR_ADDR_TO}"
         fi
+        DEBUG_STR="$CUR_CMD ${CUR_DEBUG_TO}"
         ;;
     write|copy)
         get_next_lexeme
@@ -383,8 +410,10 @@ while true; do
             CUR_ADDR_FROM="${CONSTANTS_ADDRESSES[$CUR_INDEX]}"
             CUR_DEBUG_FROM="${CONSTANTS[$CUR_INDEX]}"
         else
-            CUR_ADDR_FROM=$(get_address_for_lexeme "${LEXEME}")
-            CUR_DEBUG_FROM=$(get_debug_address_for_lexeme "${LEXEME}")
+            get_address_for_lexeme "${LEXEME}"
+            CUR_ADDR_FROM="$FUNC_RESULT"
+            get_debug_address_for_lexeme "${LEXEME}"
+            CUR_DEBUG_FROM="$FUNC_RESULT"
         fi
 
         get_next_lexeme
@@ -393,10 +422,13 @@ while true; do
         fi
 
         get_next_lexeme
-        CUR_ADDR_TO=$(get_address_for_lexeme "${LEXEME}")
-        CUR_DEBUG_TO=$(get_debug_address_for_lexeme "${LEXEME}")
+        get_address_for_lexeme "${LEXEME}"
+        CUR_ADDR_TO="$FUNC_RESULT"
+        get_debug_address_for_lexeme "${LEXEME}"
+        CUR_DEBUG_TO="$FUNC_RESULT"
 
-        RES_STR="$INSTR_COPY_FROM_TO_ADDRESS ${CUR_ADDR_FROM} ${CUR_ADDR_TO} # ${CUR_DEBUG_FROM} => ${CUR_DEBUG_TO}"
+        RES_STR="$INSTR_COPY_FROM_TO_ADDRESS ${CUR_ADDR_FROM} ${CUR_ADDR_TO}"
+        DEBUG_STR="${CUR_DEBUG_FROM} => ${CUR_DEBUG_TO}"
         ;;
     *)
         CUR_ERROR="Unsupported command $CUR_CMD"
@@ -404,30 +436,33 @@ while true; do
         ;;
     esac
     if [ -n "$CUR_ERROR" ]; then
-        compilation_error "$CUR_ERROR"
-        while [ -n "$LEXEME" ] && [ "$LEXEME" != "end" ]; do
-            get_next_lexeme
-        done
+        compilation_error "" "$CUR_ERROR"
+        echo "COMPILATION ERROR" >> "$KERNEL_FILE"
     fi
     if [ -n "$RES_STR" ]; then
-        echo "$RES_STR" >> "$KERNEL_FILE"
+        if [ -n "$DEBUG_STR" ] && [ "$DEBUG_INFO" -eq 1 ]; then
+            DEBUG_STR=" # $DEBUG_STR"
+        else
+            DEBUG_STR=""
+        fi
+        echo "$RES_STR$DEBUG_STR" >> "$KERNEL_FILE"
     fi
 done
 
 # Let's align all the # comments to the right:
-awk '{
-    # Find the position of the last number before #
-    pos = index($0, "#")
-    if (pos == 0) {
-        # No # found, print line as is
-        print
-    } else {
-        # Calculate spaces needed to align # to column 20
-        spaces = 15 - (pos - 1)
-        # Print the line up to # with calculated padding
-        printf "%s%*s%s\n", substr($0, 1, pos - 1), spaces, "", substr($0, pos)
-    }
-}' "${KERNEL_FILE}" > "${KERNEL_FILE}".tmp && mv "${KERNEL_FILE}".tmp "${KERNEL_FILE}"
+if [ -s "$KERNEL_FILE" ]; then
+    awk '{
+        pos = index($0, "#")
+        if (pos == 0) {
+            print
+        } else {
+            spaces = 15 - (pos - 1)
+            printf "%s%*s%s\n", substr($0, 1, pos - 1), spaces, "", substr($0, pos)
+        }
+    }' "${KERNEL_FILE}" > "${KERNEL_FILE}".tmp && mv "${KERNEL_FILE}".tmp "${KERNEL_FILE}"
+else
+    compilation_error "" "Empty kernel file: no valid instructions present in the provided source files"
+fi
 
 # Append constants after instructions
 for CUR_CONST in "${CONSTANTS_EVALUATED[@]}"; do
@@ -443,7 +478,7 @@ done
 # That's it - compilation finished. Let's report status to the user:
 if [ $COMPILATION_ERROR_COUNT -ne 0 ]; then
     echo -e "\033[91mCompilation failed: $COMPILATION_ERROR_COUNT error(s).\033[0m"
-    exit 1
+    exit $COMPILATION_ERROR_COUNT
 else
     echo -e "\033[92mCompilation succeeded. Kernel image: ${KERNEL_FILE}\033[0m"
 fi
